+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch query executions
 - Fetch query execution errors
 - Fetch query executions and execution errors by query id and query execution id
+- CSV Separator Replacement
 
 ### Changed
 - Move FHIR Packages to dktk-exporter